address pr review feedback
- add weights to evolved frame SELECT query
- preserve ascii art whitespace (strip newlines only, not spaces)
- use timingSafeEqual for session hmac comparison
- guard session token parsing with try/catch
- fail fast if SESSION_SECRET is unset
- normalize experience summary aggregates (coalesce nulls, cast count)
- validate experience endpoint input shape and numeric fields
- use x-vercel-cron header for cron auth instead of bearer token
- check profileResponse.ok before parsing github/google profiles
- set secure cookie only in production (secure: !dev)
- cast frame_count to int in creature query
- fix generationcount vs frames.length in generation status ui
- cache attractor trail at 2hz instead of recomputing every raf

+++ src/routes/api/creature/+server.ts
@@ -21,7 +21,7 @@ export const GET: RequestHandler = async ({ cookies }) => {
 
   const { rows } = await sql`
     SELECT c.id, c.created_at, c.generation_count, c.next_generation_at,
-           COUNT(f.id) as frame_count
+           COUNT(f.id)::int as frame_count
     FROM creatures c
     LEFT JOIN frames f ON f.creature_id = c.id
     WHERE c.user_id = ${session.userId}
